SailPoint Coding Challenge
Using the language of your choice, write code that will use the GitHub API to retrieve a summary of all opened, closed, and in progress pull requests in the last week for a given repository and print an email summary report that might be sent to a manager or Scrum-master. Choose any public target GitHub repository you like that has had at least 3 pull requests in the last week. Format the content email as you see fit, with the goal to allow the reader to easily digest the events of the past week. Please print to console the details of the email you would send (From, To, Subject, Body). As part of the submission, you are welcome to create a Dockerfile to build an image that will run the program, however, other ways of implementing this is acceptable.

Push your solution to a fresh GitHub or BitBucket repository and share the repository. Reply to this email with a link to the repo and your GitHub or BitBucket username once done.

Definition of Done:
- Your code demonstrates use of variables, looping structures, and control structures
- Your code prints a user-friendly summary of open, merged, and closed pull requests including counts of PRs as well as their titles. Use your judgement on what you think is important information.
- Your code is able to be run/compiled without errors
- Your code should be configurable when it runs. Use your judgement on what needs to be configurable.
- Your solution describes how this would be run to produce regular reports.

During the next interview, we ask that you share your screen and demonstrate the solution. We will also talk through your code and the design considerations you took in the implementation.

I look forward to seeing what you come up with, good luck!

Files:
------

requirements.dist
* required PIP modules to be installed
* pip install --upgrade -r requirements.dist
* pip freeze > requirements.pip

requirements.pip
* Freeze of PIP modules use in development
* pip install -r requirements.pip

api.py
* Low level API wrapper around requests

github.py
* GitHub API interface, the GitHub API logic

dotenv.py
* Supports a "vault" so API token is not stored in code
* key=value pairs go into a .env file, that is not stored in the repo

pull_requests.py
* Interface to the API json data for a single Pull Request
* Add paramenter functions for ease of access

report.py
* Main report generation file
* See report.py --help